refactor(build): carve transport-free //src:engine; relocate headers out of include/
Phase B groundwork for sipi_serve_image. The FFI seam (//src/ffi) must drive the
image engine, but sipi_lib already deps //src/ffi, so the engine cannot live in
sipi_lib without a dependency cycle. Carve a transport-free //src:engine and move
its headers out of include/ to co-locate with their sources (ADR-0003).

- //src:engine — SipiImage, the four format handlers, iiifparser, and the
  serve-time admission primitives (cache, rate limiter, memory budget). Carved
  out of sipi_lib's glob; sipi_lib now deps :engine. The engine source is free of
  the shttps HTTP transport (no Connection/Server include); the residual
  transitive shttps link edge through metadata/observability is pre-existing
  (DEV-6353), not new.
- Relocate the 18 engine headers include/ -> src/ (co-located, ADR-0003);
  include/ now holds only the server/config headers (SipiConf, SipiLua,
  SipiReport, favicon) and the ICC/generated assets.
- Repackage SipiImage above metadata: invert the lone metadata->SipiImage edge.
  Icc::iccFormatter now takes the pixel shape (bps/nc/photo) by value, the
  PhotometricInterpretation enum moves to metadata/photometric_interpretation.h,
  and `friend class Icc` is dropped. metadata no longer depends on SipiImage, so
  :engine owns SipiImage.h.
- Move urldecode from transport/Connection to shttps/util — a pure string
  utility, not HTTP transport — so the engine's IIIF identifier parser is
  transport-free. The co-located urldecode_test moves with it.
- Decouple observability from iiifparser: format_type_to_string moves to
  SipiQualityFormat (with the enum it maps); sentry.h no longer includes an
  image-engine header.

Pure relocation, no behaviour change: unit + approval (byte-identical) + e2e green.
