+++ src/SipiCommon.h
