+++ src/SipiFilenameHash.h
