+++ src/formats/SipiIOJpeg.h
