+++ src/formats/SipiIOTiff.h
