+++ src/formats/output_sink.h
