+++ src/iiifparser/SipiDecodeDims.h
