+++ src/iiifparser/SipiRegion.h
