+++ src/iiifparser/SipiRotation.h
