+++ src/iiifparser/SipiSize.h
