+++ src/shttps/BUILD.bazel
@@ -158,14 +158,13 @@ cc_library(
     ],
 )
 
-# Co-located unit tests for the transport module (ADR-0003): Connection's
-# urldecode + the SocketControl/ThreadControl pair. Depends only on the
-# transport+lua core archive, not the whole `:sipi_lib`.
+# Co-located unit tests for the transport module (ADR-0003): the
+# SocketControl/ThreadControl pair. Depends only on the transport+lua core
+# archive, not the whole `:sipi_lib`. (urldecode moved to //src/shttps/util.)
 cc_test(
     name = "transport_test",
     srcs = [
         "transport/socketcontrol_test.cpp",
-        "transport/urldecode_test.cpp",
     ],
     deps = [
         ":shttps",
