+++ test/unit/cache/cache.cpp
@@ -5,8 +5,8 @@
 
 #include "gtest/gtest.h"
 
-#include "../../../include/SipiCache.h"
-#include "../../../src/SipiError.h"
+#include "SipiCache.h"
+#include "SipiError.h"
 
 #include <cstdio>
 #include <cstring>
